Create and fetch manifest
Persistance
===========

Added Redis to act as a "persistance" layer.  Because we don't need long
term storage and Azure offers a hosted Redis cache, that will act as the
persistance layer.  Most interactions will only last as long as a
session, so Redis should handle that requirement well.

Also added a route to fetch the manifest once it has been created.